+++ fe/src/main/java/org/apache/doris/alter/RollupJobV2.java
@@ -32,6 +32,7 @@ import org.apache.doris.catalog.Tablet;
 import org.apache.doris.catalog.TabletInvertedIndex;
 import org.apache.doris.catalog.TabletMeta;
 import org.apache.doris.common.Config;
+import org.apache.doris.common.FeConstants;
 import org.apache.doris.common.MarkedCountDownLatch;
 import org.apache.doris.common.io.Text;
 import org.apache.doris.common.util.TimeUtils;
@@ -204,35 +205,37 @@ public class RollupJobV2 extends AlterJobV2 {
             db.readUnlock();
         }
 
-        // send all tasks and wait them finished
-        AgentTaskQueue.addBatchTask(batchTask);
-        AgentTaskExecutor.submit(batchTask);
-        long timeout = Math.min(Config.tablet_create_timeout_second * 1000L * totalReplicaNum,
-                Config.max_create_table_timeout_second * 1000L);
-        boolean ok = false;
-        try {
-            ok = countDownLatch.await(timeout, TimeUnit.MILLISECONDS);
-        } catch (InterruptedException e) {
-            LOG.warn("InterruptedException: ", e);
-            ok = false;
-        }
-
-        if (!ok || !countDownLatch.getStatus().ok()) {
-            // create rollup replicas failed. just cancel the job
-            // clear tasks and show the failed replicas to user
-            AgentTaskQueue.removeBatchTask(batchTask, TTaskType.CREATE);
-            String errMsg = null;
-            if (!countDownLatch.getStatus().ok()) {
-                errMsg = countDownLatch.getStatus().getErrorMsg();
-            } else {
-                List<Entry<Long, Long>> unfinishedMarks = countDownLatch.getLeftMarks();
-                // only show at most 3 results
-                List<Entry<Long, Long>> subList = unfinishedMarks.subList(0, Math.min(unfinishedMarks.size(), 3));
-                errMsg = "Error replicas:" + Joiner.on(", ").join(subList);
+        if (!FeConstants.runningUnitTest) {
+            // send all tasks and wait them finished
+            AgentTaskQueue.addBatchTask(batchTask);
+            AgentTaskExecutor.submit(batchTask);
+            long timeout = Math.min(Config.tablet_create_timeout_second * 1000L * totalReplicaNum,
+                    Config.max_create_table_timeout_second * 1000L);
+            boolean ok = false;
+            try {
+                ok = countDownLatch.await(timeout, TimeUnit.MILLISECONDS);
+            } catch (InterruptedException e) {
+                LOG.warn("InterruptedException: ", e);
+                ok = false;
             }
-            LOG.warn("failed to create rollup replicas for job: {}, {}", jobId, errMsg);
-            cancelImpl("Create rollup replicas failed. Error: " + errMsg);
-            return;
+
+            if (!ok || !countDownLatch.getStatus().ok()) {
+                // create rollup replicas failed. just cancel the job
+                // clear tasks and show the failed replicas to user
+                AgentTaskQueue.removeBatchTask(batchTask, TTaskType.CREATE);
+                String errMsg = null;
+                if (!countDownLatch.getStatus().ok()) {
+                    errMsg = countDownLatch.getStatus().getErrorMsg();
+                } else {
+                    List<Entry<Long, Long>> unfinishedMarks = countDownLatch.getLeftMarks();
+                    // only show at most 3 results
+                    List<Entry<Long, Long>> subList = unfinishedMarks.subList(0, Math.min(unfinishedMarks.size(), 3));
+                    errMsg = "Error replicas:" + Joiner.on(", ").join(subList);
+                }
+                LOG.warn("failed to create rollup replicas for job: {}, {}", jobId, errMsg);
+                cancelImpl("Create rollup replicas failed. Error: " + errMsg);
+                return;
+            }
         }
 
         // create all rollup replicas success.
