+++ fe/src/test/java/org/apache/doris/alter/SchemaChangeJobV2Test.java
@@ -1,0 +1,199 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package org.apache.doris.alter;
+
+import static org.junit.Assert.assertEquals;
+
+import org.apache.doris.alter.AlterJobV2.JobState;
+import org.apache.doris.analysis.AccessTestUtil;
+import org.apache.doris.analysis.AddColumnClause;
+import org.apache.doris.analysis.AlterClause;
+import org.apache.doris.analysis.Analyzer;
+import org.apache.doris.analysis.ColumnDef;
+import org.apache.doris.analysis.ColumnDef.DefaultValue;
+import org.apache.doris.analysis.ColumnPosition;
+import org.apache.doris.analysis.TypeDef;
+import org.apache.doris.catalog.AggregateType;
+import org.apache.doris.catalog.Catalog;
+import org.apache.doris.catalog.CatalogTestUtil;
+import org.apache.doris.catalog.Database;
+import org.apache.doris.catalog.FakeCatalog;
+import org.apache.doris.catalog.FakeEditLog;
+import org.apache.doris.catalog.MaterializedIndex;
+import org.apache.doris.catalog.MaterializedIndex.IndexExtState;
+import org.apache.doris.catalog.MaterializedIndex.IndexState;
+import org.apache.doris.catalog.OlapTable;
+import org.apache.doris.catalog.OlapTable.OlapTableState;
+import org.apache.doris.catalog.Partition;
+import org.apache.doris.catalog.Partition.PartitionState;
+import org.apache.doris.catalog.PrimitiveType;
+import org.apache.doris.catalog.Replica;
+import org.apache.doris.catalog.ScalarType;
+import org.apache.doris.catalog.Tablet;
+import org.apache.doris.common.AnalysisException;
+import org.apache.doris.common.FeConstants;
+import org.apache.doris.common.FeMetaVersion;
+import org.apache.doris.common.UserException;
+import org.apache.doris.meta.MetaContext;
+import org.apache.doris.task.AgentTask;
+import org.apache.doris.task.AgentTaskQueue;
+import org.apache.doris.thrift.TTaskType;
+import org.apache.doris.transaction.FakeTransactionIDGenerator;
+import org.apache.doris.transaction.GlobalTransactionMgr;
+
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
+
+import java.lang.reflect.InvocationTargetException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+/*
+ * Author: Chenmingyu
+ * Date: Aug 21, 2019
+ */
+
+public class SchemaChangeJobV2Test {
+
+    private static FakeEditLog fakeEditLog;
+    private static FakeCatalog fakeCatalog;
+    private static FakeTransactionIDGenerator fakeTransactionIDGenerator;
+    private static GlobalTransactionMgr masterTransMgr;
+    private static GlobalTransactionMgr slaveTransMgr;
+    private static Catalog masterCatalog;
+    private static Catalog slaveCatalog;
+
+    private static Analyzer analyzer;
+    private static ColumnDef newCol = new ColumnDef("add_v", new TypeDef(ScalarType.createType(PrimitiveType.INT)),
+            false, AggregateType.MAX, false, new DefaultValue(true, "1"), "");
+    private static AddColumnClause addColumnClause = new AddColumnClause(newCol, new ColumnPosition("v"), null, null);
+
+    @Before
+    public void setUp() throws InstantiationException, IllegalAccessException, IllegalArgumentException,
+            InvocationTargetException, NoSuchMethodException, SecurityException, AnalysisException {
+        fakeEditLog = new FakeEditLog();
+        fakeCatalog = new FakeCatalog();
+        fakeTransactionIDGenerator = new FakeTransactionIDGenerator();
+        masterCatalog = CatalogTestUtil.createTestCatalog();
+        slaveCatalog = CatalogTestUtil.createTestCatalog();
+        MetaContext metaContext = new MetaContext();
+        metaContext.setMetaVersion(FeMetaVersion.VERSION_60);
+        metaContext.setThreadLocalInfo();
+
+        masterTransMgr = masterCatalog.getGlobalTransactionMgr();
+        masterTransMgr.setEditLog(masterCatalog.getEditLog());
+        slaveTransMgr = slaveCatalog.getGlobalTransactionMgr();
+        slaveTransMgr.setEditLog(slaveCatalog.getEditLog());
+        analyzer = AccessTestUtil.fetchAdminAnalyzer(false);
+        addColumnClause.analyze(analyzer);
+
+        FeConstants.runningUnitTest = true;
+    }
+
+    @Test
+    public void testAddSchemaChange() throws UserException {
+        FakeCatalog.setCatalog(masterCatalog);
+        SchemaChangeHandler schemaChangeHandler = Catalog.getInstance().getSchemaChangeHandler();
+        ArrayList<AlterClause> alterClauses = new ArrayList<>();
+        alterClauses.add(addColumnClause);
+        Database db = masterCatalog.getDb(CatalogTestUtil.testDbId1);
+        OlapTable olapTable = (OlapTable) db.getTable(CatalogTestUtil.testTableId1);
+        schemaChangeHandler.process(alterClauses, "default_cluster", db, olapTable);
+        Map<Long, AlterJobV2> alterJobsV2 = schemaChangeHandler.getAlterJobsV2();
+        Assert.assertEquals(1, alterJobsV2.size());
+        Assert.assertEquals(OlapTableState.SCHEMA_CHANGE, olapTable.getState());
+    }
+
+    // start a schema change, then finished
+    @Test
+    public void testSchemaChange1() throws Exception {
+        FakeCatalog.setCatalog(masterCatalog);
+        SchemaChangeHandler schemaChangeHandler = Catalog.getInstance().getSchemaChangeHandler();
+
+        // add a schema change job
+        ArrayList<AlterClause> alterClauses = new ArrayList<>();
+        alterClauses.add(addColumnClause);
+        Database db = masterCatalog.getDb(CatalogTestUtil.testDbId1);
+        OlapTable olapTable = (OlapTable) db.getTable(CatalogTestUtil.testTableId1);
+        Partition testPartition = olapTable.getPartition(CatalogTestUtil.testPartitionId1);
+        schemaChangeHandler.process(alterClauses, "default_cluster", db, olapTable);
+        Map<Long, AlterJobV2> alterJobsV2 = schemaChangeHandler.getAlterJobsV2();
+        Assert.assertEquals(1, alterJobsV2.size());
+        SchemaChangeJobV2 schemaChangeJob = (SchemaChangeJobV2) alterJobsV2.values().stream().findAny().get();
+
+        MaterializedIndex baseIndex = testPartition.getBaseIndex();
+        assertEquals(IndexState.NORMAL, baseIndex.getState());
+        assertEquals(PartitionState.NORMAL, testPartition.getState());
+        assertEquals(OlapTableState.SCHEMA_CHANGE, olapTable.getState());
+
+        Tablet baseTablet = baseIndex.getTablets().get(0);
+        List<Replica> replicas = baseTablet.getReplicas();
+        Replica replica1 = replicas.get(0);
+        Replica replica2 = replicas.get(1);
+        Replica replica3 = replicas.get(2);
+
+        assertEquals(CatalogTestUtil.testStartVersion, replica1.getVersion());
+        assertEquals(CatalogTestUtil.testStartVersion, replica2.getVersion());
+        assertEquals(CatalogTestUtil.testStartVersion, replica3.getVersion());
+        assertEquals(-1, replica1.getLastFailedVersion());
+        assertEquals(-1, replica2.getLastFailedVersion());
+        assertEquals(-1, replica3.getLastFailedVersion());
+        assertEquals(CatalogTestUtil.testStartVersion, replica1.getLastSuccessVersion());
+        assertEquals(CatalogTestUtil.testStartVersion, replica2.getLastSuccessVersion());
+        assertEquals(CatalogTestUtil.testStartVersion, replica3.getLastSuccessVersion());
+
+        // runPendingJob
+        schemaChangeHandler.runOneCycle();
+        Assert.assertEquals(JobState.WAITING_TXN, schemaChangeJob.getJobState());
+        Assert.assertEquals(2, testPartition.getMaterializedIndices(IndexExtState.ALL).size());
+        Assert.assertEquals(1, testPartition.getMaterializedIndices(IndexExtState.VISIBLE).size());
+        Assert.assertEquals(1, testPartition.getMaterializedIndices(IndexExtState.SHADOW).size());
+        
+        // runWaitingTxnJob
+        schemaChangeHandler.runOneCycle();
+        Assert.assertEquals(JobState.RUNNING, schemaChangeJob.getJobState());
+
+        // runWaitingTxnJob, task not finished
+        schemaChangeHandler.runOneCycle();
+        Assert.assertEquals(JobState.RUNNING, schemaChangeJob.getJobState());
+
+        // runRunningJob
+        schemaChangeHandler.runOneCycle();
+        // task not finished, still running
+        Assert.assertEquals(JobState.RUNNING, schemaChangeJob.getJobState());
+
+        // finish alter tasks
+        List<AgentTask> tasks = AgentTaskQueue.getTask(TTaskType.ALTER);
+        Assert.assertEquals(3, tasks.size());
+        for (AgentTask agentTask : tasks) {
+            agentTask.setFinished(true);
+        }
+        MaterializedIndex shadowIndex = testPartition.getMaterializedIndices(IndexExtState.SHADOW).get(0);
+        for (Tablet shadowTablet : shadowIndex.getTablets()) {
+            for (Replica shadowReplica : shadowTablet.getReplicas()) {
+                shadowReplica.updateVersionInfo(testPartition.getVisibleVersion(), testPartition.getVisibleVersionHash(), shadowReplica.getDataSize(), shadowReplica.getRowCount());
+            }
+        }
+        
+        schemaChangeHandler.runOneCycle();
+        Assert.assertEquals(JobState.FINISHED, schemaChangeJob.getJobState());
+    }
+
+}
